Change width and height to doubles

+++ src/main/java/breakout/Paddle.java
@@ -26,7 +26,7 @@ public class Paddle extends Sprite {
    * @param position Top left position of the Paddle.
    * @param color Color for the fill of the Rectangle.
    */
-  public Paddle(int width, int height, Point2D position, Color color) {
+  public Paddle(double width, double height, Point2D position, Color color) {
     super(new Rectangle(width, height, color),
         position,
         new Point2D(0, 0));
